add avatar component
