+++ src/components/Avatar/index.ts
@@ -1,1 +1,1 @@
-export { Avatar } from './Avatar';
+export * from './Avatar';
