+++ src/legacy/index.ts
@@ -1,3 +1,4 @@
+export * from './Avatar';
 export * from './Button';
 export * from './Checkbox';
 export * from './Switch';
